Clean up dpnp ufunc code.
